@@ -167,23 +167,24 @@ class PageMetrics extends FixedScrollMetrics {
     required double? viewportDimension,
     required AxisDirection axisDirection,
     required this.viewportFraction,
+    required double devicePixelRatio,
   }) : super(
-          minScrollExtent: minScrollExtent,
-          maxScrollExtent: maxScrollExtent,
-          pixels: pixels,
-          viewportDimension: viewportDimension,
-          axisDirection: axisDirection,
-        );
+            minScrollExtent: minScrollExtent,
+            maxScrollExtent: maxScrollExtent,
+            pixels: pixels,
+            viewportDimension: viewportDimension,
+            axisDirection: axisDirection,
+            devicePixelRatio: devicePixelRatio);
 
   @override
-  PageMetrics copyWith({
-    double? minScrollExtent,
-    double? maxScrollExtent,
-    double? pixels,
-    double? viewportDimension,
-    AxisDirection? axisDirection,
-    double? viewportFraction,
-  }) {
+  PageMetrics copyWith(
+      {double? minScrollExtent,
+      double? maxScrollExtent,
+      double? pixels,
+      double? viewportDimension,
+      AxisDirection? axisDirection,
+      double? viewportFraction,
+      double? devicePixelRatio}) {
     return PageMetrics(
       minScrollExtent: minScrollExtent ?? this.minScrollExtent,
       maxScrollExtent: maxScrollExtent ?? this.maxScrollExtent,
@@ -191,6 +192,7 @@ class PageMetrics extends FixedScrollMetrics {
       viewportDimension: viewportDimension ?? this.viewportDimension,
       axisDirection: axisDirection ?? this.axisDirection,
       viewportFraction: viewportFraction ?? this.viewportFraction,
+      devicePixelRatio: devicePixelRatio ?? this.devicePixelRatio,
     );
   }
 
@@ -266,7 +268,7 @@ class _PagePosition extends ScrollPositionWithSingleContext
 
   @override
   void saveScrollOffset() {
-    PageStorage.of(context.storageContext)?.writeState(
+    PageStorage.of(context.storageContext).writeState(
         context.storageContext,
         getPageFromPixels(hasPixels ? pixels : null,
             hasViewportDimension ? viewportDimension : null));
@@ -276,7 +278,7 @@ class _PagePosition extends ScrollPositionWithSingleContext
   void restoreScrollOffset() {
     if (hasPixels == true) {
       final double? value = PageStorage.of(context.storageContext)
-          ?.readState(context.storageContext);
+          .readState(context.storageContext);
       if (value != null) _pageToUseOnStartup = value;
     }
   }
@@ -306,6 +308,7 @@ class _PagePosition extends ScrollPositionWithSingleContext
     double? viewportDimension,
     AxisDirection? axisDirection,
     double? viewportFraction,
+    double? devicePixelRatio,
   }) {
     return PageMetrics(
       minScrollExtent: minScrollExtent ??
@@ -317,6 +320,7 @@ class _PagePosition extends ScrollPositionWithSingleContext
           ((hasViewportDimension) ? this.viewportDimension : null),
       axisDirection: axisDirection ?? this.axisDirection,
       viewportFraction: viewportFraction ?? this.viewportFraction,
+      devicePixelRatio: devicePixelRatio ?? this.devicePixelRatio,
     );
   }
 }
